--compile warning
more get rid of data type converting warnings on windows.

+++ src/lottie/lottieitem.cpp
@@ -156,15 +156,15 @@ const LOTLayerNode *LOTCompItem::renderTree() const
 
 bool LOTCompItem::render(const rlottie::Surface &surface)
 {
-    VBitmap bitmap(reinterpret_cast<uchar *>(surface.buffer()), surface.width(),
-                   surface.height(), surface.bytesPerLine(),
+    VBitmap bitmap(reinterpret_cast<uchar *>(surface.buffer()),
+                   uint(surface.width()), uint(surface.height()), uint(surface.bytesPerLine()),
                    VBitmap::Format::ARGB32_Premultiplied);
 
     /* schedule all preprocess task for this frame at once.
      */
     mDrawableList.clear();
     mRootLayer->renderList(mDrawableList);
-    VRect clip(0, 0, surface.drawRegionWidth(), surface.drawRegionHeight());
+    VRect clip(0, 0, int(surface.drawRegionWidth()), int(surface.drawRegionHeight()));
     for (auto &e : mDrawableList) {
         e->preprocess(clip);
     }
@@ -172,8 +172,8 @@ bool LOTCompItem::render(const rlottie::Surface &surface)
     VPainter painter(&bitmap);
     // set sub surface area for drawing.
     painter.setDrawRegion(
-        VRect(surface.drawRegionPosX(), surface.drawRegionPosY(),
-              surface.drawRegionWidth(), surface.drawRegionHeight()));
+        VRect(int(surface.drawRegionPosX()), int(surface.drawRegionPosY()),
+              int(surface.drawRegionWidth()), int(surface.drawRegionHeight())));
     mRootLayer->render(&painter, {}, {});
 
     return true;
